Remove the NOT NULL value here
The newer version of mariadb that I'm using apparently is in strict mode.
This column of the table has no default value, and doesn't want things to
be null. I'm updating it so that it does have a default of null.

This makes sense cuz we use LDAP for all credentialing.